최영민 - 회원 탈퇴 갤러리 데이터 및 커플 테이블 삭제 완료

+++ macaron/src/main/java/kr/spring/gallery/service/GalleryServiceImpl.java
@@ -111,9 +111,11 @@ public class GalleryServiceImpl implements GalleryService{
 	@Override
 	public void deleteGallery(String g_cookie) {
 		int[] g_numList = galleryMapper.selectGnumForDelete(g_cookie);
-		galleryMapper.deleteGalleryImg(g_numList);
-		galleryMapper.deleteAllGalleryReply(g_numList);
-		galleryMapper.deleteGallery(g_cookie);
+		if(g_numList.length != 0) {
+			galleryMapper.deleteGalleryImg(g_numList);
+			galleryMapper.deleteAllGalleryReply(g_numList);
+			galleryMapper.deleteGallery(g_cookie);			
+		}
 	}
